Architecture change, merge server and client, proper message emit
