+++ example_config.py
@@ -1,3 +1,4 @@
 ip_address = '127.0.0.1'
 port = 1303
 server_quality = 200  # View README to see which quality to choose
+
